+++ CHANGES.md
@@ -23,6 +23,7 @@
 * Remove `Helper.normalize`.
 * Remove `Linear::DSL.insert_task`. The canonical way to add steps is using the ADDS interface going through a normalizer.
   This is why there's a normalizer for `end` (or "terminus") now for consistency.
+* In Normalizer, `path.wirings` is now `activity.wirings`.
 
 # 0.5.0